NBLAST a single flycircuit neuron against database and return scores

NBLAST a single flycircuit neuron against database and return scores

flycircuit_gmr_topn finds the top GMR Gal4 lines matching a FlyCircuit neuron query.

flycircuit_topn(query, n = 50, ...)

flycircuit_gmr_topn(query, n = 50)

Arguments

query
A flycircuit identifier

n
The number of scores to return (if n=0 or n is greater than the number of scores, then all scores are returned.

...
Additional arguments passed to flycircuit_nblast

Value

a data.frame in descending score order with columns

id Flycircuit Neuron name (or GMR id for flycircuit_gmr_topn)

score NBLAST score (normalised -1 (bad) to +1 (perfect)). Scores above 0 are potentially interesting, scores above about 0.4 seriously interesing.

Examples

# NOT RUN {
## example of remote call via curl

## simple version collecting json results directly (RECOMMENDED)
curl http://vfbdev.inf.ed.ac.uk/ocpu/library/flynblastscores/R/flycircuit_topn/json \
  -d 'query="FruMARCM-M002262_seg001"'
# top 3 hits only
curl http://vfbdev.inf.ed.ac.uk/ocpu/library/flynblastscores/R/flycircuit_topn/json \
  -d 'query="FruMARCM-M002262_seg001"&n=3'

## more flexible version
curl http://vfbdev.inf.ed.ac.uk/ocpu/library/flynblastscores/R/flycircuit_topn \
  -d 'query="FruMARCM-M002262_seg001"'
# (return value looks like this and includes sesion key)
/ocpu/tmp/x05f62b2975/R/.val
/ocpu/tmp/x05f62b2975/messages
/ocpu/tmp/x05f62b2975/stdout
/ocpu/tmp/x05f62b2975/source
/ocpu/tmp/x05f62b2975/console
/ocpu/tmp/x05f62b2975/info
/ocpu/tmp/x05f62b2975/files/DESCRIPTION

# now collect results
curl http://vfbdev.inf.ed.ac.uk/ocpu/tmp/x05f62b2975/R/.val/json
curl http://vfbdev.inf.ed.ac.uk/ocpu/tmp/x05f62b2975/R/.val/csv

# specify number of hits, use "neuron" identifier
curl http://vfbdev.inf.ed.ac.uk/ocpu/library/flynblastscores/R/flycircuit_topn \
  -d 'query="fru-M-200266"&n=10'
# }

# NBLAST a flycircuit neuron
qid="VGlut-F-100184"
res=flycircuit_gmr_topn(qid)
res

# return all results
resall=flycircuit_gmr_topn(qid, n=0)
nrow(resall)

# look at co-registered top hits on VFB
if(require('vfbr')){
 # find vfb ids for hits
 vfbids=vfbr::gmr_vfbid(res$id)
 # find vfb id for query
 qvfbid=vfb_tovfbids(qid)
 # stack browser URL for query neuron and top 10 hits
 u=vfb_stack_url(rev(c(qvfbid, vfbids[1:10])), clear = TRUE)
 

#> Error: <text>:19:0: unexpected end of input
#> 17:  u=vfb_stack_url(rev(c(qvfbid, vfbids[1:10])), clear = TRUE)
#> 18:  
#>    ^
# NOT RUN {
 browseURL(u)

# }
}

#> Error: <text>:1:1: unexpected '}'
#> 1: }
#>     ^